% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evid_map.R
\name{light_grey}
\alias{light_grey}
\title{Define 'light grey' for 0 in evidence gap maps}
\usage{
light_grey()
}
\value{
Hard coded to be grey90.
}
\description{
Define 'light grey' for 0 in evidence gap maps
}
