% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evid_map.R
\name{plot_bar_col}
\alias{plot_bar_col}
\title{Plot bar plot on the top/bottom of an evidence gap map for column frequency.}
\usage{
plot_bar_col(
  var_col,
  position = "top",
  max_freq = NULL,
  fill = "steelblue",
  margins = c(0.5, 0, 4, 4)
)
}
\arguments{
\item{var_col}{The variable to be plotted as columns.}

\item{max_freq}{Maximum frequency corresponding to \code{color_range[2]}.
Default is \code{NULL}, i.e., based on data. Can also choose a larger
integer value. Useful when plotting multiple evidence maps and want to
ensure the same color brightness refers to the same frequency.}

\item{fill}{Filling color for the bar plot. Default is \code{"steelblue"}.}

\item{margins}{Margins for the evidence gap map, in the order of bottom,
left, top, right (see \code{\link[grid]{plotViewport}}).}
}
\description{
Plot bar plot on the top/bottom of an evidence gap map for column frequency.
}
