% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evid_map.R
\name{plot_bar_row}
\alias{plot_bar_row}
\title{Plot bar plot on the left of an evidence gap map for row frequency.}
\usage{
plot_bar_row(
  var_row,
  var_row_levels = NULL,
  fill = "steelblue",
  color_zero = "grey90",
  margins = c(0.5, 0, 4, 4)
)
}
\arguments{
\item{var_row}{The variable to be plotted as rows.}

\item{var_row_levels}{Ordering of levels in \code{var_row}. Can include
values not in \code{var_row_levels} (in which case 0 will be plotted). The
first level will be plotted in the top row. Default is \code{NULL}, i.e.,
all unique values in \code{var_row} in alphabetic order.}

\item{fill}{Filling color for the bar plot. Default is \code{"steelblue"}.}

\item{margins}{Margins for the evidence gap map, in the order of bottom,
left, top, right (see \code{\link[grid]{plotViewport}}).}
}
\description{
Plot bar plot on the left of an evidence gap map for row frequency.
}
