% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evid_map.R
\name{plot_evid_map}
\alias{plot_evid_map}
\title{Plot an evidence gap map for frequency based on two variables}
\usage{
plot_evid_map(
  var_row,
  var_col,
  var_row_levels = NULL,
  print_var_row_levels = TRUE,
  position_var_col = "top",
  n_y = NULL,
  color_range = c("white", "steelblue"),
  max_freq = NULL,
  color_zero = "grey90",
  fill_zero = NULL,
  margins = c(0.5, 4, 4, 0.5)
)
}
\arguments{
\item{var_row}{The variable to be plotted as rows.}

\item{var_col}{The variable to be plotted as columns.}

\item{var_row_levels}{Ordering of levels in \code{var_row}. Can include
values not in \code{var_row_levels} (in which case 0 will be plotted). The
first level will be plotted in the top row. Default is \code{NULL}, i.e.,
all unique values in \code{var_row} in alphabetic order.}

\item{print_var_row_levels}{Whether to print levels in \code{var_row} on the
left of evidence gap map (default is \code{TRUE}). Useful when plotting
multiple maps side by side with the same row labels.}

\item{position_var_col}{Where to print levels of \code{var_col} (default is
\code{"top"}, can also choose \code{"bottom"}).}

\item{n_y}{Height of the evidence gap map. Default is \code{NULL}, i.e.,
based on the number of levels in \code{var_row}. Can also choose larger
integer values. Useful when plotting evidence maps of different sizes side
by side.}

\item{color_range}{A vector of color for 0 and maximum frequency in the
evidence gap map. Default is \code{c("white", "steelblue")}.}

\item{max_freq}{Maximum frequency corresponding to \code{color_range[2]}.
Default is \code{NULL}, i.e., based on data. Can also choose a larger
integer value. Useful when plotting multiple evidence maps and want to
ensure the same color brightness refers to the same frequency.}

\item{margins}{Margins for the evidence gap map, in the order of bottom,
left, top, right (see \code{\link[grid]{plotViewport}}).}
}
\value{
A plot made using the \code{grid} package.
}
\description{
Plot an evidence gap map for frequency based on two variables
}
